Decouple XRay
